add gitkeep for failed_tests_record

+++ ditorch/ditorch_test/utils/clean.py
@@ -8,6 +8,7 @@ shutil.rmtree(os.path.join(dirtorch_test_dir, "origin_torch"), ignore_errors=Tru
 shutil.rmtree(os.path.join(dirtorch_test_dir, "processed_tests"), ignore_errors=True)
 
 for filename in os.listdir(os.path.join(dirtorch_test_dir, "failed_tests_record")):
-    file_path = os.path.join(dirtorch_test_dir, "failed_tests_record", filename)
-    if os.path.isfile(file_path):
-        os.remove(file_path)
+    if filename.endswith("json"):
+        file_path = os.path.join(dirtorch_test_dir, "failed_tests_record", filename)
+        if os.path.isfile(file_path):
+            os.remove(file_path)
